getting set up
moving to python 3,  moving the views to use jinja2 templates
getting /static/ files to work on my machine.

+++ mealscountserver/csvparser.py
@@ -1,15 +1,16 @@
-from mealcountschool import mealcountschool
+from .mealcountschool import mealcountschool
+
 
 def parseSchoolInfo(csvLine):
-	components = csvLine.split(',')
-	return mealcountschool(components[0], long(components[1]), long(components[2]))
+    components = csvLine.split(',')
+    return mealcountschool(components[0], components[1], components[2])
 
 
 def parseCsv(csvString):
-	csvLines = csvString.splitlines()
-	mealCountsSchools = []
-	# Ignore first line
-	csvLines = csvLines[1:]
-	for csvLine in csvLines:
-		mealCountsSchools = [parseSchoolInfo(csvLine)] + mealCountsSchools
-	return mealCountsSchools
+    csvLines = csvString.splitlines()
+    mealCountsSchools = []
+    # Ignore first line
+    csvLines = csvLines[1:]
+    for csvLine in csvLines:
+        mealCountsSchools = [parseSchoolInfo(csvLine)] + mealCountsSchools
+    return mealCountsSchools
